app/backend/requirements.txt
django==3.2.4
pygdal==3.6.2.11
gdal==3.0.0
pytz==2024.1
sqlparse==0.4.4

django-debug-toolbar==3.8.1
psycopg2-binary==2.9.9
djangorestframework==3.15.1
drf-yasg==1.21.7
django-crispy-forms==2.0
django-formtools==2.4.1
django-filter==23.5
requests==2.27.1
whitenoise==6.5.0
django-cors-headers==4.1.0
django-extensions==3.2.3
django-rest-multiple-models==2.1.3
django-reversion==5.0.12
geopandas==0.10.2
thefuzz==0.20.0
djangorestframework-simplejwt==5.3.0
cryptography==42.0.8
geojson==3.1.0
minio==7.2.6
openpyxl==2.5.14
djangorestframework-csv==3.0.0
coverage==7.2.0
MarkupSafe==2.1.5

gunicorn==19.9.0
gevent==1.3.0
django-settings-export>=1.2.1
lxml==4.6.3
urllib3>=1.24,<1.25
pyjwt>=2.0,<=2.4.0